Finished the game, save, and load menus
Also touched up on a lot of the methods in the GameRendering class.

+++ Game Code/Cabin Escape/src/com/CabinEscape/GameSettings.java
@@ -11,10 +11,10 @@ public class GameSettings {
     //Ints
     public final int gameWindowWidth = 120;                                      //The width of the games window (amount of rows in the ASCII Terminal)
     public final int gameWindowHeight = 50;                                     //The height of the games window (amount of columns in the ASCII Terminal)
-    public final int SOUND_MAX = 10;
-    public int masterSound = SOUND_MAX;
-    public int musicSound = SOUND_MAX;
-    public int fxSound = SOUND_MAX;
+    public final double SOUND_MAX = 10;
+    public double masterSound = SOUND_MAX;
+    public double musicSound = SOUND_MAX;
+    public double fxSound = SOUND_MAX;
     
     //Rects
     public final Rect gameWindowRect = new Rect (0, 0, 120, 50);
